+++ branches/custom-train-tracks/Shared/sdk/SString.hpp
@@ -121,11 +121,11 @@ SString& SString::vFormat ( const char* szFormat, va_list vl )
 void SString::Split ( const SString& strDelim, std::vector < SString >& outResult, unsigned int uiMaxAmount, unsigned int uiMinAmount ) const
 {
     outResult.clear ();
-    unsigned long ulStartPoint = 0;
+    size_t ulStartPoint = 0;
 
     while ( true )
     {
-        unsigned long ulPos = find ( strDelim, ulStartPoint );
+        size_t ulPos = find ( strDelim, ulStartPoint );
 
         if ( ulPos == npos || ( uiMaxAmount > 0 && uiMaxAmount <= outResult.size () + 1 ) )
         {
@@ -157,7 +157,7 @@ bool SString::Split ( const SString& strDelim, SString* pstrLeft, SString* pstrR
 
     assert ( iIndex );
     bool bFromEnd = iIndex < 0;
-    unsigned long ulPos;
+    size_t ulPos;
     if ( !bFromEnd )
     {
         ulPos = 0;
@@ -391,28 +391,28 @@ SString SString::Left ( int iCount ) const
 // Right most number of characters
 SString SString::Right ( int iCount ) const
 {
-    return SubStr ( length () - iCount, iCount );
+    return SubStr ( (int)length () - iCount, iCount );
 }
 
 
 bool SString::EndsWith ( const SString& strOther ) const
 {
-    return Right ( strOther.length () ) == strOther;
+    return Right ( (int)strOther.length () ) == strOther;
 }
 
 bool SString::EndsWithI ( const SString& strOther ) const
 {
-    return stricmp ( Right ( strOther.length () ), strOther ) == 0;
+    return stricmp ( Right ( (int)strOther.length () ), strOther ) == 0;
 }
 
 bool SString::BeginsWith ( const SString& strOther ) const
 {
-    return Left ( strOther.length () ) == strOther;
+    return Left ( (int)strOther.length () ) == strOther;
 }
 
 bool SString::BeginsWithI ( const SString& strOther ) const
 {
-    return stricmp ( Left ( strOther.length () ), strOther ) == 0;
+    return stricmp ( Left ( (int)strOther.length () ), strOther ) == 0;
 }
 
 // Static function
